# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blog Pessoal",
    "version": "1.0.0",
    "description": "API RESTful completa para um sistema de blog pessoal desenvolvida em Laravel com autenticação JWT como trabalho de conclusão da matéria Sistemas Distribuídos e API da PÓS em FullStack da Unimar.",
    "contact": {
      "name": "Fabiano Souza de Oliveira",
      "email": "fsouzaweb@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autenticação JWT. Use: Bearer {seu_token_jwt}"
      }
    },
    "schemas": {
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Laravel"},
          "slug": {"type": "string", "example": "laravel"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
...
  "paths": {
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Listar todas as tags",
        "description": "Retorna uma lista de tags, com opção de listar as mais populares",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "popular",
            "in": "query",
            "description": "Listar apenas tags populares",
            "schema": {"type": "boolean"}
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limite de tags populares (padrão: 10)",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tags obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Tag"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Criar nova tag",
        "description": "Cria uma nova tag no sistema",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "Vue.js"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Tag"}
                  }
                }
              }
            }
          },
          "422": {"description": "Dados de validação inválidos"}
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": ["Tags"],
        "summary": "Exibir tag específica",
        "description": "Retorna os dados de uma tag específica",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da tag",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Tag encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Tag"}
                  }
                }
              }
            }
          },
          "404": {"description": "Tag não encontrada"}
        }
      },
      "put": {
        "tags": ["Tags"],
        "summary": "Atualizar tag",
        "description": "Atualiza os dados de uma tag existente",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da tag",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "React.js"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Tag"}
                  }
                }
              }
            }
          },
          "404": {"description": "Tag não encontrada"},
          "422": {"description": "Dados de validação inválidos"}
        }
      },
      "delete": {
        "tags": ["Tags"],
        "summary": "Excluir tag",
        "description": "Remove uma tag do sistema",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da tag",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Tag excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "Tag excluída com sucesso"}
                  }
                }
              }
            }
          },
          "404": {"description": "Tag não encontrada"}
        }
      }
    }
  }
}